Proper speed mining implementation. Turns out spamming the move order makes a BIG difference (not sure why). The first roach comes out ~6 seconds earlier now!

+++ Bot/GameData/Buildings.cs
@@ -7,7 +7,7 @@ public static class Buildings {
     public static float GetRadius(uint buildingType) {
         var buildingAbilityId = KnowledgeBase.GetUnitTypeData(buildingType).AbilityId;
 
-        return KnowledgeBase.GetAbilityData((int)buildingAbilityId).FootprintRadius;
+        return KnowledgeBase.GetAbilityData(buildingAbilityId).FootprintRadius;
     }
 
     public class Dimension {
